add boot2docker ip to NO_PROXY list
The docker client respects the HTTP_PROXY environment variable for api calls since version 1.5.

This kind of breaks the boot2docker scenario because docker is running on a local VM which most most certainly won't be reachable by the proxy.

This commit changes `boot2docker shellinit` to append the DOCKER_HOST to the NO_PROXY variable.