Fix shellinit output for cmd.exe/powershell
Also fixes the issue with Windows paths with backslashes in the bash
`export` statements.

When shellinit/up executed on windows, prints instructions on how to set
environment variables correctly for Windows Command Prompt (cmd.exe) and
PowerShell. (only the powershell set path statements are printed to
stdout; the rest goes to stderr).

Signed-off-by: Ahmet Alp Balkan <ahmetalpbalkan@gmail.com>